+++ packages/google_maps_flutter/google_maps_flutter_web/CHANGELOG.md
@@ -1,3 +1,8 @@
+
+## 0.3.3
+
+* Adds options for gesture handling and tilt controls.
+
 ## 0.3.2+1
 
 * Removes dependency on `meta`.